fixing 14

+++ 14-binary_tree_balance.c
@@ -8,7 +8,7 @@
 
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int ans = 0;
+	size_t ans = 0;
 
 	if (tree == NULL)
 		return (0);
